CMakeLists.txt: introduce 2 option variables, and make Skia usable on non-Apple platforms.
* SKIA_LIBRARY_PATH
A variable to specify the full-path location of the skia
library to be linked. If you want to link SKIA.DLL, you
should specify "SKIA.DLL.LIB", the export symbol file
instead of runtime library.

* WIN32_OBJECT_TYPE
A variable to specify the sharable/non-sharable compiler
switch for MSVC (or Clang-Cl). You should choose "/MD"
(for sharable object) or "/MT" (for non-sharable object).
This flag is not used on non-Windows platforms.