This is a fix for the string parser code overstepping to "end()"
When parsing strings, the code sometimes oversteps the string iterator to end or past end and then dereferences without checking if it's valid to do so.
This results in an exception being thrown in newer versions of Visual Studio.

The change involves moving the code from passing a current position and end iterator to various functions. Instead, they're passed an object that tracks both of those and ensures that:
1. The iterator never steps past end
2. The iterator is only dereferenced when it is safe to do so

This change also includes some cleanup of warnings on Windows